+++ Changelog.md
@@ -5,6 +5,7 @@ Bug fixes
 * Restore `--full_backtrace` option
 * Ensure that values passed to `config.filter_run` are respected when running
   over DRb (using spork).
+* Ensure shared example groups are reset after a run (as example groups are).
 
 ### 2.8.0 / 2012-01-04